Updated IndexSequence to support long for sequence length and updated SequenceGenerator::sequence to accept long for the length of the generated sequence. Updates to Expression implementations, and some other types, are made for compatibility with this and ExpressionMatrix::create is updated to decide whether it is feasible to create an ExpressionMatrix and return null if it is not.

+++ code/src/main/java/io/almostrealism/expression/BooleanConstant.java
@@ -41,7 +41,7 @@ public class BooleanConstant extends Constant<Boolean> {
 	}
 
 	@Override
-	public IndexSequence sequence(Index index, int len) {
+	public IndexSequence sequence(Index index, long len) {
 		return IndexSequence.of(value ? 1 : 0, len);
 	}
