+++ code/src/main/java/io/almostrealism/expression/LongConstant.java
@@ -70,7 +70,7 @@ public class LongConstant extends Constant<Long> {
 	public Number value(IndexValues indexValues) { return value; }
 
 	@Override
-	public IndexSequence sequence(Index index, int len) {
+	public IndexSequence sequence(Index index, long len) {
 		return IndexSequence.of(value, len);
 	}
